refactor: extract pane-agnostic diff-core from merge editor
The merge editor's rendering machinery was written for exactly three panes.
A read-only diff viewer needs the same layout math, ribbon geometry, scroll
sync, line numbers, and syntax highlighting for two, so the parts that never
depended on there being a middle pane move to src/webviews/react/diff-core/
and now take an ordered pane-id array instead of left/middle/right.

The vertical layout engine, the adjacent-pair ribbon paths, the scroll
drivers, the generic block shells, and the shared stylesheet are all pane-
count agnostic; merge-only pieces stay behind, including bandSpansForMiddleGap
and every accept/reject action. merge-editor.css imports diff-core.css and
keeps its own variables, which the core reads through --diff-* aliases so a
second bundle can supply its own.

Contract tests cover a two-pane layout and a three-pane one, asserting whole
ribbon paths rather than a shared substring so the two cases cannot pass on
the same bytes.

Behavior is unchanged: same DOM, same class names, same protocol, and all 32
pixel baselines pass without re-recording.
