+++ src/webviews/react/merge-editor/segments.tsx
@@ -1,370 +1,25 @@
-// Merge editor segment rendering components.
-// Each pane (ours/result/theirs) renders in its own column, so a segment is
-// split into per-pane blocks: CommonPaneBlock for unchanged code and the
-// Ours/Result/Theirs ConflictBlock trio for a hunk (with resolution controls and
-// an editable result). ConnectorLayer draws the ribbons linking a hunk across
-// panes; OverviewRail provides a minimap of conflict locations for navigation.
+// Merge-only conflict rendering components.
+// Generic code-line rendering, syntax highlighting, line numbers, and common
+// segment shells live in diff-core; resolution controls stay in this module.
 
-import React, { useCallback, useMemo, useState } from "react";
-import type { CommonSegment, ConflictSegment, HunkResolution, HunkSideDismissal } from "./types";
-import {
-    tokenSimilarityRatio,
-    buildWordDiffMask,
-    tokenizeWordDiff,
-    alignCompareLinesForWordDiff,
-} from "../../../mergeEditor/wordDiff";
+import React, { useCallback, useState } from "react";
+import type { ConflictSegment, HunkResolution, HunkSideDismissal } from "./types";
 import { getEffectiveResultLines, splitEditedText } from "./mergeState";
-import { tokenizeSyntaxLine, type SyntaxTokenKind } from "./syntaxHighlight";
-import { highlightLine } from "./shikiHighlighter";
-import { useSyntaxHighlightState, type SyntaxHighlightState } from "./syntaxHighlightContext";
-import type { LineNumberValue } from "./lineNumbers";
-import { LINE_HEIGHT_PX, type MergePane } from "./mergeScrollLayout";
 import { t } from "../shared/i18n";
+import {
+    CodeBlock,
+    intrinsicSizeStyle,
+    lineNumberSpecEqual,
+    LineNumbers,
+    type LineNumberSpec,
+} from "../diff-core/segments";
 
-// --- Syntax highlighting ---
-
-const TOKEN_CLASS: Record<SyntaxTokenKind, string | undefined> = {
-    plain: undefined,
-    comment: "tok-comment",
-    string: "tok-string",
-    keyword: "tok-keyword",
-    constant: "tok-constant",
-    number: "tok-number",
-};
-
-/** A single colored run of text: Shiki tokens carry an inline style, fallback tokens carry a className. */
-interface ColoredSpan {
-    text: string;
-    style?: React.CSSProperties;
-    className?: string;
+/** Line-number specifications for the merge editor's three protocol panes. */
+export interface SegmentPaneLineNumbers {
+    left: LineNumberSpec;
+    middle: LineNumberSpec;
+    right: LineNumberSpec;
 }
-
-// Shiki `fontStyle` bitmask bits (see @shikijs/core ThemedToken).
-const FONT_STYLE_ITALIC = 1;
-const FONT_STYLE_BOLD = 2;
-const FONT_STYLE_UNDERLINE = 4;
-
-/**
- * Tokenizes a line into colored spans. Uses Shiki's grammar-accurate tokens
- * when the highlighter is ready and the file's language is registered;
- * otherwise falls back to the hand-rolled five-category tokenizer.
- */
-function coloredSpansForLine(line: string, ctx: SyntaxHighlightState): ColoredSpan[] {
-    if (ctx.ready && ctx.lang) {
-        const shikiTokens = highlightLine(line, ctx.lang, ctx.theme);
-        if (shikiTokens) {
-            return shikiTokens.map((tok) => {
-                const style: React.CSSProperties = {};
-                if (tok.color) style.color = tok.color;
-                if (tok.fontStyle) {
-                    if (tok.fontStyle & FONT_STYLE_ITALIC) style.fontStyle = "italic";
-                    if (tok.fontStyle & FONT_STYLE_BOLD) style.fontWeight = "bold";
-                    if (tok.fontStyle & FONT_STYLE_UNDERLINE) style.textDecoration = "underline";
-                }
-                return { text: tok.text, style: Object.keys(style).length ? style : undefined };
-            });
-        }
-    }
-    return tokenizeSyntaxLine(line).map((token) => ({
-        text: token.text,
-        className: TOKEN_CLASS[token.kind],
-    }));
-}
-
-function renderColoredSpans(spans: ColoredSpan[], keyPrefix: string): React.ReactNode[] {
-    let offset = 0;
-    return spans.map((span) => {
-        const key = `${keyPrefix}-${offset}-${span.text}`;
-        offset += span.text.length;
-        return (
-            <span key={key} className={span.className} style={span.style}>
-                {span.text}
-            </span>
-        );
-    });
-}
-
-const HighlightedLine = React.memo(function HighlightedLine({
-    line,
-}: {
-    line: string;
-}): React.ReactElement {
-    const ctx = useSyntaxHighlightState();
-    if (!line) return <>{` `}</>;
-    // Pure syntax-token helper, not a component invocation.
-    // react-doctor-disable-next-line react-doctor/no-render-in-render
-    return <>{renderColoredSpans(coloredSpansForLine(line, ctx), "line")}</>;
-});
-
-/**
- * Expands a token-level word-diff mask into per-character changed/whitespace
- * masks aligned to `line`. This lets the change overlay be intersected with
- * Shiki's full-line colored spans without re-tokenizing fragments (which
- * would lose surrounding grammar context, e.g. a string opened before the
- * changed run).
- */
-function buildChangedCharMasks(
-    line: string,
-    compareLine: string,
-): { changed: boolean[]; whitespace: boolean[] } {
-    const tokens = tokenizeWordDiff(line);
-    const changedMask = buildWordDiffMask(line, compareLine);
-    const changed: boolean[] = [];
-    const whitespace: boolean[] = [];
-    for (let i = 0; i < tokens.length; i++) {
-        const token = tokens[i];
-        const isChanged = changedMask[i];
-        const isWhitespace = /^\s+$/.test(token);
-        for (let c = 0; c < token.length; c++) {
-            changed.push(isChanged);
-            whitespace.push(isWhitespace);
-        }
-    }
-    return { changed, whitespace };
-}
-
-/**
- * Renders colored spans with a word-diff overlay: each colored span is split
- * at change-boundary offsets so the underlying grammar coloring is preserved
- * beneath the `.word-diff-change` background instead of being replaced by it.
- */
-function renderColoredSpansWithWordDiff(
-    spans: ColoredSpan[],
-    changed: boolean[],
-    whitespace: boolean[],
-    keyPrefix: string,
-): React.ReactNode[] {
-    const nodes: React.ReactNode[] = [];
-    let offset = 0;
-    for (const span of spans) {
-        const spanText = span.text;
-        let runStart = 0;
-        while (runStart < spanText.length) {
-            const runChanged = changed[offset + runStart];
-            let runEnd = runStart + 1;
-            while (runEnd < spanText.length && changed[offset + runEnd] === runChanged) {
-                runEnd++;
-            }
-            const runText = spanText.slice(runStart, runEnd);
-            const key = `${keyPrefix}-${offset + runStart}`;
-            const coloredNode = (
-                <span key={key} className={span.className} style={span.style}>
-                    {runText}
-                </span>
-            );
-            if (runChanged) {
-                const runIsWhitespace = whitespace[offset + runStart];
-                nodes.push(
-                    <span
-                        key={`chg-${key}`}
-                        className={`word-diff-change ${runIsWhitespace ? "word-diff-whitespace" : ""}`}
-                    >
-                        {coloredNode}
-                    </span>,
-                );
-            } else {
-                nodes.push(coloredNode);
-            }
-            runStart = runEnd;
-        }
-        offset += spanText.length;
-    }
-    return nodes;
-}
-
-const WordDiffLine = React.memo(function WordDiffLine({
-    line,
-    compareLine,
-}: {
-    line: string;
-    compareLine: string;
-}): React.ReactElement {
-    const ctx = useSyntaxHighlightState();
-    if (!line) return <>{` `}</>;
-    if (line === compareLine) return <HighlightedLine line={line} />;
-    if (!compareLine) return <HighlightedLine line={line} />;
-
-    const similarity = tokenSimilarityRatio(line, compareLine);
-    if (similarity < 0.28) {
-        return <HighlightedLine line={line} />;
-    }
-
-    const spans = coloredSpansForLine(line, ctx);
-    if (spans.length === 0) return <>{` `}</>;
-
-    const { changed, whitespace } = buildChangedCharMasks(line, compareLine);
-
-    const wordDiffNodes = renderColoredSpansWithWordDiff(spans, changed, whitespace, "wd");
-    return <>{wordDiffNodes}</>;
-});
-
-// --- Line numbers ---
-
-/** Line-number values to render alongside a code block (optional secondary column). */
-export interface LineNumberSpec {
-    primary: LineNumberValue[];
-    secondary?: LineNumberValue[];
-}
-
-interface LineNumbersProps extends LineNumberSpec {
-    rowIsReal?: boolean[];
-}
-
-function padLines(lines: string[], count: number): string[] {
-    const padded = [...lines];
-    while (padded.length < count) padded.push("");
-    return padded;
-}
-
-function lineNumberValuesEqual(a: LineNumberValue[], b: LineNumberValue[]): boolean {
-    if (a === b) return true;
-    if (a.length !== b.length) return false;
-    for (let i = 0; i < a.length; i++) {
-        if (a[i] !== b[i]) return false;
-    }
-    return true;
-}
-
-function lineNumberSpecEqual(a: LineNumberSpec, b: LineNumberSpec): boolean {
-    if (a === b) return true;
-    if (!lineNumberValuesEqual(a.primary, b.primary)) return false;
-    if (a.secondary === b.secondary) return true;
-    if (!a.secondary || !b.secondary) return false;
-    return lineNumberValuesEqual(a.secondary, b.secondary);
-}
-
-function rowPresenceEqual(a: boolean[] | undefined, b: boolean[] | undefined): boolean {
-    if (a === b) return true;
-    if (!a || !b) return false;
-    if (a.length !== b.length) return false;
-    for (let i = 0; i < a.length; i++) {
-        if (a[i] !== b[i]) return false;
-    }
-    return true;
-}
-
-const LineNumbers = React.memo(
-    function LineNumbers({ primary, rowIsReal }: LineNumbersProps) {
-        return (
-            <div className="line-numbers">
-                {Array.from({ length: primary.length }, (_, i) => {
-                    const isReal = rowIsReal?.[i] ?? true;
-                    return (
-                        <div
-                            key={i}
-                            className={`line-number-row ${
-                                isReal ? "real-line-row" : "padding-line-row"
-                            }`}
-                        >
-                            <div className="line-number line-number-primary">
-                                {primary[i] ?? ""}
-                            </div>
-                        </div>
-                    );
-                })}
-            </div>
-        );
-    },
-    (prev, next) =>
-        lineNumberSpecEqual(prev, next) && rowPresenceEqual(prev.rowIsReal, next.rowIsReal),
-);
-
-// --- Code block ---
-
-interface CodeBlockProps {
-    lines: string[];
-    lineCount: number;
-    lineNumbers: LineNumberSpec;
-    showLineNumbers?: boolean;
-    lineNumberSide?: "left" | "right";
-    className?: string;
-    wordHighlight?: boolean;
-    compareLines?: string[];
-}
-
-function rowKey(lineNumbers: LineNumberSpec, line: string, row: number): string {
-    const primary = lineNumbers.primary[row] ?? "gap";
-    const secondary = lineNumbers.secondary?.[row] ?? "gap";
-    return `${primary}-${secondary}-${row}-${line}`;
-}
-
-const CodeBlock = React.memo(
-    function CodeBlock({
-        lines,
-        lineCount,
-        lineNumbers,
-        showLineNumbers = true,
-        lineNumberSide = "left",
-        className,
-        wordHighlight,
-        compareLines,
-    }: CodeBlockProps) {
-        // Padding rows align panes to the tallest side; only source-backed rows
-        // should receive diff coloring.
-        const rowCount = Math.max(lineCount, lines.length);
-        const rowIsReal = useMemo(
-            () => Array.from({ length: rowCount }, (_, i) => i < lines.length),
-            [lines.length, rowCount],
-        );
-        const padded = useMemo(() => padLines(lines, rowCount), [lines, rowCount]);
-        const paddedCompare = useMemo(() => {
-            if (!compareLines) return undefined;
-            const alignedCompare = alignCompareLinesForWordDiff(lines, compareLines);
-            return padLines(alignedCompare, rowCount);
-        }, [compareLines, lines, rowCount]);
-
-        return (
-            <div
-                className={`code-block ${showLineNumbers ? `line-numbers-${lineNumberSide}` : "no-line-numbers"} ${className ?? ""} ${wordHighlight ? "word-highlight" : ""}`}
-            >
-                {showLineNumbers && lineNumberSide === "left" ? (
-                    <LineNumbers
-                        primary={lineNumbers.primary}
-                        secondary={lineNumbers.secondary}
-                        rowIsReal={rowIsReal}
-                    />
-                ) : null}
-                <div className="code-lines">
-                    {padded.map((line, i) => {
-                        const isReal = rowIsReal[i] ?? false;
-                        return (
-                            <div
-                                key={rowKey(lineNumbers, line, i)}
-                                className={`code-line ${
-                                    isReal ? "real-code-line" : "padding-code-line"
-                                }`}
-                            >
-                                <span className="code-line-content">
-                                    {wordHighlight && paddedCompare ? (
-                                        <WordDiffLine line={line} compareLine={paddedCompare[i]} />
-                                    ) : (
-                                        <HighlightedLine line={line} />
-                                    )}
-                                </span>
-                            </div>
-                        );
-                    })}
-                </div>
-                {showLineNumbers && lineNumberSide === "right" ? (
-                    <LineNumbers
-                        primary={lineNumbers.primary}
-                        secondary={lineNumbers.secondary}
-                        rowIsReal={rowIsReal}
-                    />
-                ) : null}
-            </div>
-        );
-    },
-    (prev, next) =>
-        prev.lines === next.lines &&
-        prev.lineCount === next.lineCount &&
-        prev.showLineNumbers === next.showLineNumbers &&
-        prev.lineNumberSide === next.lineNumberSide &&
-        prev.className === next.className &&
-        prev.wordHighlight === next.wordHighlight &&
-        prev.compareLines === next.compareLines &&
-        lineNumberSpecEqual(prev.lineNumbers, next.lineNumbers),
-);
 
 // --- Editable result block ---
 
@@ -459,77 +114,7 @@ function EditableResultBlock({
     );
 }
 
-// --- Hunk helpers ---
-
-/** Line-number specifications for the left, result, and right merge panes. */
-export interface SegmentPaneLineNumbers {
-    left: LineNumberSpec;
-    middle: LineNumberSpec;
-    right: LineNumberSpec;
-}
-
-// --- Section components ---
-
-/**
- * Size hint that lets `content-visibility: auto` skip layout of offscreen
- * segments while keeping the scrollbar geometry stable for large files. Every
- * block — common or conflict — is exactly `lineCount * LINE_HEIGHT_PX` tall
- * (conflict rules are drawn with a zero-height inset shadow), so this matches
- * both the rendered content box and the scroll geometry.
- */
-function intrinsicSizeStyle(lineCount: number): React.CSSProperties {
-    return { containIntrinsicSize: `auto ${lineCount * LINE_HEIGHT_PX}px` };
-}
-
-interface CommonPaneBlockProps {
-    pane: MergePane;
-    segment: CommonSegment;
-    lineCount: number;
-    lineNumbers: LineNumberSpec;
-    highlightWords: boolean;
-}
-
-/**
- * Renders one pane's slice of an unchanged segment. The three panes hold
- * identical common lines but flow in separate columns (PyCharm-style), so each
- * is its own block; the scroll driver keeps them vertically aligned. Memoized
- * with value-compared line numbers so resolving one hunk does not re-render
- * every other segment.
- */
-export const CommonPaneBlock = React.memo(
-    function CommonPaneBlock({
-        pane,
-        segment,
-        lineCount,
-        lineNumbers,
-        highlightWords,
-    }: CommonPaneBlockProps) {
-        return (
-            <div className="segment segment-common" style={intrinsicSizeStyle(lineCount)}>
-                <CodeBlock
-                    lines={segment.lines}
-                    lineCount={lineCount}
-                    lineNumbers={lineNumbers}
-                    lineNumberSide={pane === "left" ? "right" : "left"}
-                    wordHighlight={highlightWords}
-                />
-            </div>
-        );
-    },
-    (prev, next) =>
-        prev.pane === next.pane &&
-        prev.segment === next.segment &&
-        prev.lineCount === next.lineCount &&
-        prev.highlightWords === next.highlightWords &&
-        lineNumberSpecEqual(prev.lineNumbers, next.lineNumbers),
-);
-
-/**
- * Props shared by all three conflict-pane blocks: result-line computation
- * inputs, selection, active-state styling, and word highlighting. `editedLines`
- * overrides the side-resolution result when set. Side blocks add resolution
- * callbacks; the result block adds the edit callback and ordinals.
- */
+/** Props shared by all merge conflict pane blocks. */
 export interface ConflictPaneBaseProps {
     segment: ConflictSegment;
     resolution: HunkResolution | undefined;
